backend/requirements.txt
requests
python-dotenv
elevenlabs
flask
flask-cors